add compree class

+++ lw3/streams/DecompressInputStreamDecorator.cpp
@@ -36,7 +36,7 @@ streamsize CDecompressInputStreamDecorator::ReadBlock(void * dstBuffer, streamsi
 
 void CDecompressInputStreamDecorator::UpdateBuffer(std::streamsize size)
 {
-    while (m_buffer.size() < size)
+    while (m_buffer.size() < static_cast<size_t>(size))
     {
         Chunk chunk;
         m_baseStream->ReadBlock(reinterpret_cast<void*>(&chunk), 2);
